backend/src/api/requirements.txt
alembic==1.17.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.4
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
postgrest==2.27.0
propcache==0.4.1
protobuf==6.33.2
psycopg==3.3.2
psycopg2==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
realtime==2.27.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
sib-api-v3-sdk==7.6.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
starlette==0.50.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
